% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_attendance}
\alias{check_attendance}
\title{check_attendance()}
\usage{
check_attendance(log = "upload_log.txt", att_file = "attendance_upload.csv")
}
\arguments{
\item{log}{character vector: the name of the file containing the upload log from ProActive. Can be a full or relative path to the .txt file.}

\item{eto_file}{character vector: the name of the file containing attendance data from ETO. Can be a full or relative path to the .csv file.}
}
\value{
data frame
}
\description{
This function takes two files: 1) ProActive attendance upload log (text file), and 2) "[LAYC CA] DCPCSB attendance report" as inputs, and return a list of attendance records that were not processed correctly during the upload.
}
\examples{
check_attendance(log = './data/upload_log.txt', att_file = './data/attendance_upload.csv')
}

