% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check_pro_dmg}
\alias{check_pro_dmg}
\title{check_pro_dmg()}
\usage{
check_pro_dmg(pro_file = "proactive.csv", eto_file = "eto_usi.csv")
}
\arguments{
\item{pro_file}{character vector: the name of the file containing data from ProActive. Can be a full or relative path to the .csv file.}

\item{eto_file}{character vector: the name of the file containing data from ETO. Can be a full or relative path to the .csv file.}
}
\value{
data frame
}
\description{
This function takes two .csv files: 1) ProActive full report, and 2) "[LAYC CA] Students usi report" as inputs, and return a list of records that are not in sync.
}
\examples{
check_pro_dmg(pro_file = './data/proactive.csv', eto_file = './data/eto.csv')
}

